fixed mediator and indicatorconfig addresses

+++ speed-dreams/source-2.2.3/src/simulatedDrivingAssistance/frontend/IndicatorConfig.cpp
@@ -184,7 +184,7 @@ IndicatorConfig* IndicatorConfig::GetInstance()
     std::ifstream file(filepath);
     getline(file, pointerName);
     file.close();
-    int pointerValue = stoi(pointerName, nullptr, 16);
+    long pointerValue = std::stol(pointerName, nullptr, 16);
     m_instance = (IndicatorConfig*)pointerValue;
     return m_instance;
 }
